requirements.txt
Django==6.0.2
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
